update traffic sim graph
